renamed controllers.js to userController.js
